\name{aws.sqs-package}
\alias{aws.sqs-package}
\alias{aws.sqs}
\docType{package}
\title{AWS SQS Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Simple Queue Service (SQS) REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
