% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{delete_msg}
\alias{delete_msg}
\title{delete_msg}
\usage{
delete_msg(queue, handle, ...)
}
\arguments{
\item{queue}{A character string containing a queue URL, or the name of the queue.}

\item{handle}{A message handle, as returned by \code{\link{receive_msg}}.}

\item{...}{Additional arguments passed to \code{\link{sqsHTTP}}.}
}
\value{
If operation succeeds, a logical \code{TRUE}.
}
\description{
Delete one or more messages from an SQS queue
}
\details{
Delete one or more messages from an SQS queue. If a message is not deleted, it remains visible in the queue and will be returned by subsequent calls to \code{\link{receive_msg}}.
}
\author{
Thomas J. Leeper
}
\references{
\href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_DeleteMessage.html}{DeleteMessage}

\href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_DeleteMessageBatch.html}{DeleteMessageBatch}
}
\seealso{
\code{link{receive_msg}}
}

