% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.sqs-package.R
\docType{package}
\name{aws.sqs-package}
\alias{aws.sqs}
\alias{aws.sqs-package}
\title{AWS SQS Client Package}
\description{
A simple client package for the Amazon Web Services (AWS) Simple Queue
Service (SQS) REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

