% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_visibility.R
\name{visibility}
\alias{visibility}
\title{Change Visiblity Timeout}
\usage{
visibility(queue, handle, timeout = NULL, query = NULL, ...)
}
\arguments{
\item{queue}{A character string containing a queue URL, or the name of the queue.}

\item{handle}{A character vector containing one or more message handles, as returned by \code{\link{receive_msg}}.}

\item{timeout}{An integer value indicating the new value of the visibility timeout, in seconds between 0 and 43200, for the message(s).}

\item{query}{A list specifying additional query arguments to be passed to the \code{query} argument of \code{\link{sqsHTTP}}.}

\item{...}{Additional arguments passed to \code{\link{sqsHTTP}}.}
}
\value{
If successful, a logical \code{TRUE}. Otherwise, a data structure of class \dQuote{aws_error} containing any error message(s) from AWS and information about the request attempt.
}
\description{
Change the visibility timeout of received message(s)
}
\details{
Changes the visibility timeout of one or more SQS messages, by their message handle.
}
\references{
\href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_ChangeMessageVisibility.html}{ChangeMessageVisibility}
}
\seealso{
\code{link{receive_msg}} \code{link{delete_msg}}
}
\author{
Thomas J. Leeper
}
